'cluster' is now factored out into a separate library, 'Muster', so it's removed from the Libra build.  We'll add support for sampled tracing back later with a configure option to build with Muster.
Changed the way Libra builds with MPI.  Now the mpi compiler is autodetected and
includes and libraries are extracted from it by running it with -show, -showme, etc.  This allows
the MPI support to integrate better into an autoconf build, as we can just treat MPI like any
other library.  Previous we had to build *everything* with the MPI compiler, which resulted
in unnecessary MPI dependencies for binaries that didn't depend on MPI.